DaoUser created UserRepository created Added reading from DataBase when we try to LogIn.
fixed the problem which only one user could log in.
now you can signUp any number of users!

To continue:
The User Details are NOT encrypted.

ERRORS & BUGS:
1. When the UserName && Password DO NOT MATCH, the app crashing
2. Different accounts have the same TABLES.
User A can use Users B tables and info!

SOLUTIONS
1. if statements..
2. Add to the creation of the table "DataBaseClient2" + user.getFirstName.
the FirstName is primary key so we can use it to build unique TABLES