DaoUser created UserRepository created Added reading from DataBase when we try to LogIn.
To continue:
The User Details are NOT encrypted.

FIXED:
App dosent crash if the user dosent exists.

ERRORS & BUGS:
1. Different accounts have the same TABLES.
User A can use Users B tables and info!
2. the behavior of the app when there is no user name in the table and a user name with wrond password, is the same!

SOLUTIONS
1. Add to the creation of the table "DataBaseClient2" + user.getFirstName.
the FirstName is primary key so we can use it to build unique TABLES
2. tried to fix with flag.
dosent work.
maybe need to consider to change the DaoUser QUERY.
BUT! i think its ok that they behave theSame!